// rss-css-selectors/src/components/view/popup/enum.ts
enum STYLES {
  PopupContainer = 'popup-container',
  Popup = 'popup-win',
  BtnClose = 'btn-close',
  PopupOpen = 'popup-container--open',
}

enum TAGNAMES {
  Div = 'div',
}

enum CONTENT {
  Popup = 'Congratulations!!!\n You are good at music!',
}

enum EVENTS {
  Click = 'click',
}

export { STYLES, CONTENT, TAGNAMES, EVENTS };